+++ hyperdbg/hprdbghv/header/common/State.h
@@ -332,6 +332,8 @@ typedef struct _VIRTUAL_MACHINE_STATE
     UINT8                   LastExceptionOccuredInHost;                         // The vector of last exception occured in host
     UINT64                  HostIdt;                                            // host Interrupt Descriptor Table (actual type is SEGMENT_DESCRIPTOR_INTERRUPT_GATE_64*)
     UINT64                  HostGdt;                                            // host Global Descriptor Table (actual type is SEGMENT_DESCRIPTOR_32* or SEGMENT_DESCRIPTOR_64*)
+    UINT64                  HostTss;                                            // host Task State Segment (actual type is TASK_STATE_SEGMENT_64*)
+    UINT64                  HostInterruptStack;                                 // host interrupt RSP
 
     //
     // EPT Descriptors